pyproject.toml
[project]
name = "gitrel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Eivind Fonn", email = "evfonn@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.8",
    "httpx>=0.28.1",
    "inquirer>=3.4.0",
    "pydantic>=2.10.6",
    "pygithub>=2.6.1",
    "semantic-version>=2.10.0",
    "tqdm>=4.67.1",
    "xdg-base-dirs>=6.0.2",
]

[project.scripts]
gitrel = "gitrel:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pytest>=8.3.5",
    "ruff>=0.11.0",
]

[tool.mypy]
plugins = ["pydantic.mypy"]
files = ["src/**/*.py", "tests/**/*.py"]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
show_error_codes = true

[tool.ruff]
line-length = 110
include = ["src/**/*.py", "tests/**/*.py"]

[tool.ruff.lint]
select = [
    "F",        # Pyflakes rules
    "W",        # PyCodeStyle warnings
    "E",        # PyCodeStyle errors
    "I",        # Sort imports properly
    "UP",       # Warn if certain things can changed due to newer Python versions
    "C4",       # Catch incorrect use of comprehensions, dict, list, etc
    "FA",       # Enforce from __future__ import annotations
    "ISC",      # Good use of string concatenation
    "ICN",      # Use common import conventions
    "RET",      # Good return practices
    "SIM",      # Common simplification rules
    "TID",      # Some good import practices
    "TCH",      # Enforce importing certain types in a TYPE_CHECKING block
    "PTH",      # Use pathlib instead of os.path
    "TD",       # Be diligent with TODO comments
    "NPY",      # Some numpy-specific things
]

[tool.pytest.ini_options]
testpaths = ["tests"]
